Kreiranje tabele Uloga i redizajniranje baze @T2539:R

+++ BioskopJPA/src/model/Sifarnik.java
@@ -2,7 +2,6 @@ package model;
 
 import java.io.Serializable;
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.util.List;
 
 
@@ -19,7 +18,7 @@ public class Sifarnik implements Serializable {
 	@GeneratedValue(strategy=GenerationType.IDENTITY)
 	private int sifarnikID;
 
-	private BigDecimal cena;
+	private double cena;
 
 	private String tip;
 
@@ -38,11 +37,11 @@ public class Sifarnik implements Serializable {
 		this.sifarnikID = sifarnikID;
 	}
 
-	public BigDecimal getCena() {
+	public double getCena() {
 		return this.cena;
 	}
 
-	public void setCena(BigDecimal cena) {
+	public void setCena(double cena) {
 		this.cena = cena;
 	}
